scripts/core/iso_manager/requirements.txt
pyyaml>=6.0
requests>=2.31.0
rich>=13.7.0